% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_to_coord}
\alias{id_to_coord}
\title{Convert Linearized Coordinates Into x/y/z}
\usage{
id_to_coord(mesh, map, normalize = FALSE)
}
\arguments{
\item{mesh}{list of numeric matrices as produced by \code{\link[=rtini_extract]{rtini_extract()}}.}

\item{map}{numeric matrix elevation map used to produce the mesh.}

\item{normalize}{FALSE unimplemented}
}
\value{
list with three components, one for each dimension.  Each component
is a three row matrix in which each column represents one triangle.
}
\description{
\code{\link[=rtini_extract]{rtini_extract()}} produces "linearized" coordinates that are useful for
directly indexing back into the elevation map, but can be clumsy for
plotting.  "Linearized" coordinates are offsets that can be used to access a
matrix or array as if it were a dimensionless vector.  This function
converts those coordinates to x/y/z values.
}
\examples{
err <- rtini_error(volcano)
mesh <- rtini_extract(err, 10)
coord <- id_to_coord(mesh, volcano)
}
