% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{rtini_error}
\alias{rtini_error}
\title{Compute RTIN Surface Approximation Error on a Height Map}
\usage{
rtini_error(map)
}
\arguments{
\item{map}{numeric matrix with at least three rows and three columns, and
with an odd number of each.  It is expected that the input is finite and
does not contain NAs.}
}
\value{
numeric matrix of the same dimension as \code{map} with the
maximum of the rtin approximation error at each point and the equivalent
approximation at the child points.
}
\description{
A vectorized adaptation of \href{https://observablehq.com/@mourner/martin-real-time-rtin-terrain-mesh}{Vladimir Agafonkin's MARTINI implementation}
of the Evans etal. \href{https://www.cs.ubc.ca/~will/papers/rtin.pdf}{Right Triangulated Irregular Networks} surface approximation
algorithm.  Accepts a broader set of input matrix dimensions than MARTINI.
}
\examples{
err <- rtini_error(volcano)
rtin_extract(err, tol=10)
}
\seealso{
\code{\link[=rtin_extract]{rtin_extract()}}
}
