% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{rtini_extract}
\alias{rtini_extract}
\title{Extract A Triangle Mesh With Approximation Error Below Tolerance}
\usage{
rtini_extract(error, tol)
}
\arguments{
\item{error}{a numeric matrix as produced by \code{\link[=rtini_error]{rtini_error()}}}

\item{tol}{scalar positive numeric, the error tolerance}
}
\value{
a list of numeric matrices.  Each column in a matrix corresponds to
the "linearized" vertex coordinates of a triangle.  See examples for how to
convert to x-y coordinates.  Each matrix corresponds to a different
approximation layer.
}
\description{
Given a surface approximation error matrix as computed with \code{\link[=rtini_error]{rtini_error()}},
extract a triangle mesh that approximates the original heightmap with error
less than or equal to a specified tolerance.
}
\examples{
err <- rtini_error(volcano)
mesh <- rtini_extract(err, tol=10)

## Convert linearized indices to x/y in [0,1]
tris <- rbind(do.call(cbind, mesh), NA)
plot.new()
polygon(
  x=((tris - 1) \%/\% nrow(err)) / (ncol(err) - 1),
  y=((tris - 1) \%\% nrow(err)) / (nrow(err) - 1),
  col='grey90'
)

## One benefit of the linearized indices is that you can
## index directly into the height and error matrices.  Here
## We use that to show the errors at every point that was not
## plotted versus volcano height.
plot(volcano[-unlist(mesh)], err[-unlist(mesh)])
}
\seealso{
\code{\link[=rtini_error]{rtini_error()}}
}
